+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-10-04 02:42:16
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 8013327963d47eec2abef6edf5e223fccf31f14d
+
+## [ForwardWidgets-2kuai] - 2025-10-04 02:42:16
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 646dafcf0f4121e977ad49202880e8d05025a0ca
+
+## [ForwardWidgets-pack1r] - 2025-10-04 02:42:16
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 9b36ab824a05b4169e7b2df53ee1ddca6780d6a8
+
+
 ## [ForwardWidgets-huangxd] - 2025-10-03 02:47:54
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main